rename timesyncbeamsearch
